fix types

+++ core/apps/api/services/payment.py
@@ -1,7 +1,8 @@
+from decimal import Decimal
 from core.apps.shared.utils.settings import energy_price
 
 
-def calc_energy_price(energy: int) -> float:
+def calc_energy_price(energy: Decimal) -> Decimal:
     """energiya narxini xisoblash
 
     Args:
